Refactor: 내가 쓴 하이라이트 목록 조회 API 반환 형식 List로 변경
- 하이라이트가 존재하지 않는 경우 예외 처리 추가
- 패키지명 오타 수정
